requirements-test.txt
pytest==3.1.1
pytest-django==3.1.2
factory-boy==2.8.1